feat(sync-notes): flush pending note debounces on disable / logout
Task 19 du plan sub-epic 03 sync-notes.

- Relocate debounce registry de useNotes vers notes-store (module non-React,
  importable depuis SyncContext sans circular dep hook → context).
- Nouvelles helpers : scheduleNoteUpdatePush, cancelNoteUpdatePush, flushPendingNoteUpdates.
- SyncContext.disableSync (déclenché aussi par auth signed-out) flush
  pendingNoteUpdates en push immédiat avant de désactiver. Évite la perte des
  2 dernières secondes de frappe au logout.
- useNotes refactorisé pour utiliser les helpers (plus de Map locale).

Tests : 257/259 (1 échec pré-existant TryItStep onboarding sans rapport).
tsc --noEmit clean.
